Suppression des fichiers de configuration et de journalisation, ajout de la logique de jeu pour le mode Light Tracker
