--- src/classes/config.py
@@ -1,19 +1,0 @@
-####################################
-# LED STRIP SETTINGS
-####################################
-GPIO_PIN = 10
-NUMBER_OF_LEDS = 176
-FREQUENCY = 800000
-DMA_CHANNEL = 10
-PLAYER_OFFSETS = {
-    1: (132, 176),
-    2: (88, 132),
-    3: (44, 88),
-    4: (0, 44)
-} 
-
-
-GREEN = (0, 255, 0)
-ORANGE = (255, 165, 0)
-RED = (255, 0, 0)
-YELLOW = (255, 255, 0)
